refactored auth components
