+++ src/components/auth/NameWindow.js
@@ -1,8 +1,8 @@
 import React, { useContext } from "react";
 import styled from "styled-components";
 import { Card as CARD, Input, Button } from "@bootstrap-styled/v4";
-import { Swatch } from "./color/Swatch.js";
-import { isNobody } from "../utils/auth.js";
+import { Swatch } from "../color/Swatch.js";
+import { isNobody } from "../../utils/auth.js";
 import { AuthContext } from "./AuthProvider.js";
 
 export const NameWindow = (props) => {
